requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
filelock==3.13.4
fsspec==2024.3.1
ftfy==6.2.0
idna==3.7
Jinja2==3.1.3
joblib==1.4.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
pillow==10.3.0
regex==2024.4.16
requests==2.31.0
scikit-learn==1.4.2
scipy==1.13.0
sympy==1.12
threadpoolctl==3.4.0
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
